show privilege count on collection option
