+++ client/src/pages/user/dialogs/types.ts
@@ -1,0 +1,41 @@
+import type { RBACOptions } from '../Types';
+
+export type Privilege = {
+  [key: string]: boolean; // key: privilege name, value: whether it's selected
+};
+
+export type CollectionPrivileges = {
+  [collectionValue: string]: Privilege; // key: collection value, value: privileges
+};
+
+export type DBPrivileges = {
+  collections: CollectionPrivileges; // Collection-level privileges
+};
+
+export type DBCollectionsPrivileges = {
+  [dbValue: string]: DBPrivileges; // key: DB value, value: DB privileges and collections
+};
+
+export type CollectionOption = {
+  name: string;
+  value: string;
+};
+
+export type DBOption = {
+  name: string;
+  value: string;
+};
+
+export interface DBCollectionsSelectorProps {
+  selected: DBCollectionsPrivileges; // Current selected DBs and their collections with privileges
+  setSelected: (
+    value:
+      | DBCollectionsPrivileges
+      | ((prev: DBCollectionsPrivileges) => DBCollectionsPrivileges)
+  ) => void;
+  // Callback to update selected state
+  options: {
+    rbacOptions: RBACOptions; // Available RBAC options (privileges)
+    dbOptions: DBOption[]; // Available databases
+  };
+}
